% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centigrados_a_faherenheit.R
\name{centigrados_a_fahrenheit}
\alias{centigrados_a_fahrenheit}
\title{Convertidor de centigrados a fahrenheit}
\usage{
centigrados_a_fahrenheit(centigrados)
}
\arguments{
\item{centigrados}{temperatura en grados centigrados}
}
\value{
Temperatura en grados fahrenheit
}
\description{
Convertidor de centigrados a fahrenheit
}
\examples{
centigrados_a_fahrenheit(25)

}
