% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fahrenheit_a_centigrados.R
\name{fahrenheit_a_centigrados}
\alias{fahrenheit_a_centigrados}
\title{Convertidor de fahrenheit a centigrados}
\usage{
fahrenheit_a_centigrados(fahrenheit)
}
\arguments{
\item{fahrenheit}{temperatura en grados fahrenheit}
}
\value{
Temperatura en grados centigrados
}
\description{
Convertidor de fahrenheit a centigrados
}
\examples{
fahrenheit_a_centigrados(50)



}
