% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafica_precipitacion.R
\name{resumen_precipitaciones_fecha}
\alias{resumen_precipitaciones_fecha}
\title{Funcion para obtener el resumen de precipitaciones por fecha}
\usage{
resumen_precipitaciones_fecha(datos)
}
\arguments{
\item{datos}{dataset que se quiere analizar}
}
\value{
Una grafica que muestra cuandos mm llovio por dia
}
\description{
Funcion para obtener el resumen de precipitaciones por fecha
}
\examples{
grafica_precipitaciones(NH0910)

}
