% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descargar_datos.R
\name{descargar_datos}
\alias{descargar_datos}
\title{Descargar datos}
\usage{
descargar_datos(id_estacion)
}
\arguments{
\item{id_estacion}{es una identificacion confomado por dos letras mayusculas y cuatro numeros}
}
\value{
Nos devuelve el dateset del id ingresado
}
\description{
Algunas de las estaciones disponibles para descargar son:
-NH0910
-NH0437
-NH0046
-NH0472
-NH0098
}
\examples{
data <- descargar datos ("NH0910")

}
